module_image_enhacement_fastapi/requirements.txt
absl-py==2.1.0
aniso8601==9.0.1
astunparse==1.6.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
plum-dispatch==1.7.4
protobuf==4.25.3
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
realesrgan-ncnn-py==2.0.0
requests==2.32.2
rich==13.7.1
setuptools==70.0.0
six==1.16.0
spire==0.4.2
Spire.Doc==12.4.0
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel
termcolor==2.4.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
urllib3==2.2.1
Wand==0.6.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
